Moves provider-test-tools tests
