+++ packages/provider-test-tools/__tests__/client-execution-order.test.js
@@ -1,5 +1,5 @@
 import gql from "graphql-tag";
-import { clientTester, getResolvedQueries } from "../";
+import { clientTester, getResolvedQueries } from "../src";
 
 function AuthorQueryResolver({ variables }) {
   return {
