+++ packages/provider-test-tools/package.json
@@ -18,13 +18,13 @@
     "testEnvironment": "jsdom",
     "coverageDirectory": "<rootDir>/packages/provider-test-tools/coverage/",
     "collectCoverageFrom": [
-      "**/packages/provider-test-tools/*.js"
+      "**/packages/provider-test-tools/src/*.js"
     ],
     "transformIgnorePatterns": [
       "node_modules/(?!(react-native|@times-components)/)"
     ],
     "testMatch": [
-      "<rootDir>/packages/provider-test-tools/src/__tests__/**.test.js"
+      "<rootDir>/packages/provider-test-tools/__tests__/**.test.js"
     ]
   },
   "repository": {
